# generated/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  license:
    name: KSTU
  title: Tourist agency
  version: 1.0.0
servers:
- url: http://127.0.0.1/v1
- url: https://fdobrotv.com/v1
paths:
...
  /cars:
    get:
      operationId: listCars
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          maximum: 100
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cars'
          description: A paged array of cars
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List all cars
      tags:
      - cars
      x-accepts: application/json
      x-tags:
      - tag: cars
    post:
      operationId: createCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarIn'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
          description: Ok
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create a car
      tags:
      - cars
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: cars
  /cars/{id}:
    delete:
      operationId: deleteCarById
      parameters:
      - description: The id of the resource to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Resource deleted successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete specific car by id
      tags:
      - cars
      x-accepts: application/json
      x-tags:
      - tag: cars
    get:
      operationId: showCarById
      parameters:
      - description: The id of the resource to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Info for a specific car
      tags:
      - cars
      x-accepts: application/json
      x-tags:
      - tag: cars
...
components:
  schemas:
...
    CarColor:
      enum:
      - white
      - black
      - brown
      - yellow
      - red
      - blue
      - silver
      type: string
...
    Cars:
      items:
        $ref: '#/components/schemas/Car'
      maxItems: 100
      type: array
    CarIn:
      example:
        markId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        color: null
        modelId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        plateNumber: A123TH716RUS
      properties:
        markId:
          format: uuid
          type: string
        modelId:
          format: uuid
          type: string
        plateNumber:
          example: A123TH716RUS
          type: string
        color:
          $ref: '#/components/schemas/CarColor'
      required:
      - color
      - markId
      - modelId
      - plateNumber
      type: object
    Car:
      example:
        color: null
        model: Traveller
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        plateNumber: A123TH716RUS
        mark: Peugeot
      properties:
        id:
          format: uuid
          type: string
        mark:
          example: Peugeot
          type: string
        model:
          example: Traveller
          type: string
        plateNumber:
          example: A123TH716RUS
          type: string
        color:
          $ref: '#/components/schemas/CarColor'
      required:
      - color
      - id
      - mark
      - model
      - plateNumber
      type: object
...
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object